Refactor form imports and improve Toggle component
- Update imports in multiple pages to use new forms directory
- Modify Toggle component to support dynamic number of options
- Add preventDefault to Toggle button click handler
- Replace select with Toggle in FormDateWithPrecision
